refactor: bigger changes like file structure, extract socket handlers
not tested well and potential bugs. Will be improved in next part
