+++ .dependabot/config.yml
@@ -2,13 +2,13 @@ version: 1
 update_configs:
   # Keep package.json (& lockfiles) up to date as soon as
   # new versions are published to the npm registry
-  - package_manager: "javascript"
-    directory: "."
-    update_schedule: "daily"
+  - package_manager: 'javascript'
+    directory: '.'
+    update_schedule: 'daily'
     automerged_updates:
       - match:
-          dependency_type: "development"
-          update_type: "semver:minor"
+          dependency_type: 'development'
+          update_type: 'semver:minor'
       - match:
-          dependency_type: "production"
-          update_type: "semver:patch"
+          dependency_type: 'production'
+          update_type: 'semver:patch'
