+++ .github/workflows/ci.yml
@@ -11,7 +11,7 @@ jobs:
       - name: Setup node
         uses: actions/setup-node@v1
         with:
-          node-version: "12"
+          node-version: '12'
       - name: Install
         run: npm ci
       - name: Check format
