+++ .prettierignore
@@ -1,2 +1,3 @@
 public/js/libs
-samples.json
+public/styles/styles.css
+*.md
