+++ docker-compose.yml
@@ -1,4 +1,4 @@
-version: "3.4"
+version: '3.4'
 
 services:
   poker:
@@ -7,7 +7,7 @@ services:
       context: .
       dockerfile: Dockerfile
     ports:
-      - "3000:3000"
+      - '3000:3000'
     environment:
       - NODE_ENV=production
-    command: "start"
+    command: 'start'
