+++ package.json
@@ -3,24 +3,26 @@
   "version": "0.0.1",
   "private": false,
   "scripts": {
-    "start": "node app.js",
-    "format": "prettier --write **/*.{js,less,json,html}",
-    "format:check": "prettier --check **/*.{js,less,json,html}"
+    "start": "node ./src/app.js",
+    "format": "prettier --write .",
+    "format:check": "prettier --check ."
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/geobalas/Poker"
   },
   "dependencies": {
+    "body-parser": "^1.19.0",
     "connect": "^3.7.0",
-    "express": "~3.4.7",
-    "express-session": "^1.17.0",
+    "express": "^4.17.1",
+    "express-session": "^1.17.1",
     "express-socket.io-session": "^1.3.5",
-    "jade": "~1.7.0",
-    "less": "~2.0.0",
-    "less-middleware": "~1.0.4",
+    "jade": "^1.11.0",
+    "less": "^3.11.1",
+    "less-middleware": "^3.1.0",
     "lodash": "^4.17.15",
-    "socket.io": "~1.2.0",
+    "morgan": "^1.10.0",
+    "socket.io": "^2.3.0",
     "winston": "^3.2.1"
   },
   "devDependencies": {
