+++ public/js/controllers/table.js
@@ -2,21 +2,21 @@
  * The table controller. It keeps track of the data on the interface,
  * depending on the replies from the server.
  */
-app.controller("TableController", [
-  "$scope",
-  "$rootScope",
-  "$http",
-  "$routeParams",
-  "$timeout",
-  "sounds",
+app.controller('TableController', [
+  '$scope',
+  '$rootScope',
+  '$http',
+  '$routeParams',
+  '$timeout',
+  'sounds',
   function ($scope, $rootScope, $http, $routeParams, $timeout, sounds) {
     var seat = null;
     $scope.table = {};
     $scope.notifications = [{}, {}, {}, {}, {}, {}, {}, {}, {}, {}];
     $scope.showingChipsModal = false;
-    $scope.actionState = "";
+    $scope.actionState = '';
     $scope.table.dealerSeat = null;
-    $scope.myCards = ["", ""];
+    $scope.myCards = ['', ''];
     $scope.mySeat = null;
     $scope.betAmount = 0;
     $scope.actionTimeout = 0;
@@ -29,8 +29,8 @@ app.controller("TableController", [
 
     // Getting the table data
     $http({
-      url: "/table-data/" + $routeParams.tableId,
-      method: "GET",
+      url: '/table-data/' + $routeParams.tableId,
+      method: 'GET',
     }).success(function (data, status, headers, config) {
       $scope.table = data.table;
       $scope.buyInAmount = data.table.maxBuyIn;
@@ -38,17 +38,17 @@ app.controller("TableController", [
     });
 
     // Joining the socket room
-    socket.emit("enterRoom", $routeParams.tableId);
+    socket.emit('enterRoom', $routeParams.tableId);
 
     $scope.minBetAmount = function () {
       if (
         $scope.mySeat === null ||
-        typeof $scope.table.seats[$scope.mySeat] === "undefined" ||
+        typeof $scope.table.seats[$scope.mySeat] === 'undefined' ||
         $scope.table.seats[$scope.mySeat] === null
       )
         return 0;
       // If the pot was raised
-      if ($scope.actionState === "actBettedPot") {
+      if ($scope.actionState === 'actBettedPot') {
         var minRaise =
           $scope.table.lastRaise < $scope.table.bigBlind
             ? $scope.table.bigBlind
@@ -68,11 +68,11 @@ app.controller("TableController", [
     $scope.maxBetAmount = function () {
       if (
         $scope.mySeat === null ||
-        typeof $scope.table.seats[$scope.mySeat] === "undefined" ||
+        typeof $scope.table.seats[$scope.mySeat] === 'undefined' ||
         $scope.table.seats[$scope.mySeat] === null
       )
         return 0;
-      return $scope.actionState === "actBettedPot"
+      return $scope.actionState === 'actBettedPot'
         ? $scope.table.seats[$scope.mySeat].chipsInPlay +
             $scope.table.seats[$scope.mySeat].bet
         : $scope.table.seats[$scope.mySeat].chipsInPlay;
@@ -81,7 +81,7 @@ app.controller("TableController", [
     $scope.callAmount = function () {
       if (
         $scope.mySeat === null ||
-        typeof $scope.table.seats[$scope.mySeat] === "undefined" ||
+        typeof $scope.table.seats[$scope.mySeat] === 'undefined' ||
         $scope.table.seats[$scope.mySeat] == null
       )
         return 0;
@@ -95,46 +95,46 @@ app.controller("TableController", [
     $scope.showLeaveTableButton = function () {
       return (
         $rootScope.sittingOnTable !== null &&
-        (!$rootScope.sittingIn || $scope.actionState === "waiting")
+        (!$rootScope.sittingIn || $scope.actionState === 'waiting')
       );
     };
 
     $scope.showPostSmallBlindButton = function () {
       return (
-        $scope.actionState === "actNotBettedPot" ||
-        $scope.actionState === "actBettedPot"
+        $scope.actionState === 'actNotBettedPot' ||
+        $scope.actionState === 'actBettedPot'
       );
     };
 
     $scope.showPostBigBlindButton = function () {
       return (
-        $scope.actionState === "actNotBettedPot" ||
-        $scope.actionState === "actBettedPot"
+        $scope.actionState === 'actNotBettedPot' ||
+        $scope.actionState === 'actBettedPot'
       );
     };
 
     $scope.showFoldButton = function () {
       return (
-        $scope.actionState === "actNotBettedPot" ||
-        $scope.actionState === "actBettedPot" ||
-        $scope.actionState === "actOthersAllIn"
+        $scope.actionState === 'actNotBettedPot' ||
+        $scope.actionState === 'actBettedPot' ||
+        $scope.actionState === 'actOthersAllIn'
       );
     };
 
     $scope.showCheckButton = function () {
       return (
-        $scope.actionState === "actNotBettedPot" ||
-        ($scope.actionState === "actBettedPot" &&
+        $scope.actionState === 'actNotBettedPot' ||
+        ($scope.actionState === 'actBettedPot' &&
           $scope.table.biggestBet == $scope.table.seats[$scope.mySeat].bet)
       );
     };
 
     $scope.showCallButton = function () {
       return (
-        $scope.actionState === "actOthersAllIn" ||
-        ($scope.actionState === "actBettedPot" &&
+        $scope.actionState === 'actOthersAllIn' ||
+        ($scope.actionState === 'actBettedPot' &&
           !(
-            $scope.actionState === "actBettedPot" &&
+            $scope.actionState === 'actBettedPot' &&
             $scope.table.biggestBet == $scope.table.seats[$scope.mySeat].bet
           ))
       );
@@ -142,7 +142,7 @@ app.controller("TableController", [
 
     $scope.showBetButton = function () {
       return (
-        $scope.actionState === "actNotBettedPot" &&
+        $scope.actionState === 'actNotBettedPot' &&
         $scope.table.seats[$scope.mySeat].chipsInPlay &&
         $scope.table.biggestBet < $scope.table.seats[$scope.mySeat].chipsInPlay
       );
@@ -150,7 +150,7 @@ app.controller("TableController", [
 
     $scope.showRaiseButton = function () {
       return (
-        $scope.actionState === "actBettedPot" &&
+        $scope.actionState === 'actBettedPot' &&
         $scope.table.seats[$scope.mySeat].chipsInPlay &&
         $scope.table.biggestBet < $scope.table.seats[$scope.mySeat].chipsInPlay
       );
@@ -158,8 +158,8 @@ app.controller("TableController", [
 
     $scope.showAllInButton = function () {
       return (
-        ($scope.actionState === "actNotBettedPot" ||
-          $scope.actionState === "actBettedPot") &&
+        ($scope.actionState === 'actNotBettedPot' ||
+          $scope.actionState === 'actBettedPot') &&
         $scope.table.seats[$scope.mySeat].chipsInPlay &&
         $scope.table.biggestBet < $scope.table.seats[$scope.mySeat].chipsInPlay
       );
@@ -167,8 +167,8 @@ app.controller("TableController", [
 
     $scope.showBetRange = function () {
       return (
-        ($scope.actionState === "actNotBettedPot" ||
-          $scope.actionState === "actBettedPot") &&
+        ($scope.actionState === 'actNotBettedPot' ||
+          $scope.actionState === 'actBettedPot') &&
         $scope.table.seats[$scope.mySeat].chipsInPlay &&
         $scope.table.biggestBet < $scope.table.seats[$scope.mySeat].chipsInPlay
       );
@@ -176,8 +176,8 @@ app.controller("TableController", [
 
     $scope.showBetInput = function () {
       return (
-        ($scope.actionState === "actNotBettedPot" ||
-          $scope.actionState === "actBettedPot") &&
+        ($scope.actionState === 'actNotBettedPot' ||
+          $scope.actionState === 'actBettedPot') &&
         $scope.table.seats[$scope.mySeat].chipsInPlay &&
         $scope.table.biggestBet < $scope.table.seats[$scope.mySeat].chipsInPlay
       );
@@ -190,15 +190,15 @@ app.controller("TableController", [
 
     $scope.potText = function () {
       if (
-        typeof $scope.table.pot !== "undefined" &&
+        typeof $scope.table.pot !== 'undefined' &&
         $scope.table.pot[0].amount
       ) {
-        var potText = "Pot: " + $scope.table.pot[0].amount;
+        var potText = 'Pot: ' + $scope.table.pot[0].amount;
 
         var potCount = $scope.table.pot.length;
         if (potCount > 1) {
           for (var i = 1; i < potCount; i++) {
-            potText += " - Sidepot: " + $scope.table.pot[i].amount;
+            potText += ' - Sidepot: ' + $scope.table.pot[i].amount;
           }
         }
         return potText;
@@ -209,23 +209,23 @@ app.controller("TableController", [
       if ($scope.mySeat === seat) {
         return $scope.myCards[card];
       } else if (
-        typeof $scope.table.seats !== "undefined" &&
-        typeof $scope.table.seats[seat] !== "undefined" &&
+        typeof $scope.table.seats !== 'undefined' &&
+        typeof $scope.table.seats[seat] !== 'undefined' &&
         $scope.table.seats[seat] &&
-        typeof $scope.table.seats[seat].cards !== "undefined" &&
-        typeof $scope.table.seats[seat].cards[card] !== "undefined"
+        typeof $scope.table.seats[seat].cards !== 'undefined' &&
+        typeof $scope.table.seats[seat].cards[card] !== 'undefined'
       ) {
-        return "card_" + $scope.table.seats[seat].cards[card];
+        return 'card_' + $scope.table.seats[seat].cards[card];
       } else {
-        return "card_back";
+        return 'card_back';
       }
     };
 
     $scope.seatOccupied = function (seat) {
       return (
         !$rootScope.sittingOnTable ||
-        ($scope.table.seats !== "undefined" &&
-          typeof $scope.table.seats[seat] !== "undefined" &&
+        ($scope.table.seats !== 'undefined' &&
+          typeof $scope.table.seats[seat] !== 'undefined' &&
           $scope.table.seats[seat] &&
           $scope.table.seats[seat].name)
       );
@@ -233,13 +233,13 @@ app.controller("TableController", [
 
     // Leaving the socket room
     $scope.leaveRoom = function () {
-      socket.emit("leaveRoom");
+      socket.emit('leaveRoom');
     };
 
     // A request to sit on a specific seat on the table
     $scope.sitOnTheTable = function () {
       socket.emit(
-        "sitOnTheTable",
+        'sitOnTheTable',
         {
           seat: selectedSeat,
           tableId: $routeParams.tableId,
@@ -252,7 +252,7 @@ app.controller("TableController", [
             $rootScope.sittingIn = true;
             $scope.buyInError = null;
             $scope.mySeat = selectedSeat;
-            $scope.actionState = "waiting";
+            $scope.actionState = 'waiting';
             $scope.$digest();
           } else {
             if (response.error) {
@@ -266,7 +266,7 @@ app.controller("TableController", [
 
     // Sit in the game
     $scope.sitIn = function () {
-      socket.emit("sitIn", function (response) {
+      socket.emit('sitIn', function (response) {
         if (response.success) {
           $rootScope.sittingIn = true;
           $rootScope.$digest();
@@ -276,12 +276,12 @@ app.controller("TableController", [
 
     // Leave the table (not the room)
     $scope.leaveTable = function () {
-      socket.emit("leaveTable", function (response) {
+      socket.emit('leaveTable', function (response) {
         if (response.success) {
           $rootScope.sittingOnTable = null;
           $rootScope.totalChips = response.totalChips;
           $rootScope.sittingIn = false;
-          $scope.actionState = "";
+          $scope.actionState = '';
           $scope.clearActionReminder();
           $rootScope.$digest();
           $scope.$digest();
@@ -291,23 +291,23 @@ app.controller("TableController", [
 
     // Post a blind (or not)
     $scope.postBlind = function (posted) {
-      socket.emit("postBlind", posted, function (response) {
+      socket.emit('postBlind', posted, function (response) {
         if (response.success && !posted) {
           $rootScope.sittingIn = false;
         } else {
           sounds.playBetSound();
         }
-        $scope.actionState = "";
+        $scope.actionState = '';
         $scope.clearActionReminder();
         $scope.$digest();
       });
     };
 
     $scope.check = function () {
-      socket.emit("check", function (response) {
+      socket.emit('check', function (response) {
         if (response.success) {
           sounds.playCheckSound();
-          $scope.actionState = "";
+          $scope.actionState = '';
           $scope.clearActionReminder();
           $scope.$digest();
         }
@@ -315,10 +315,10 @@ app.controller("TableController", [
     };
 
     $scope.fold = function () {
-      socket.emit("fold", function (response) {
+      socket.emit('fold', function (response) {
         if (response.success) {
           sounds.playFoldSound();
-          $scope.actionState = "";
+          $scope.actionState = '';
           $scope.clearActionReminder();
           $scope.$digest();
         }
@@ -326,10 +326,10 @@ app.controller("TableController", [
     };
 
     $scope.call = function () {
-      socket.emit("call", function (response) {
+      socket.emit('call', function (response) {
         if (response.success) {
           sounds.playCallSound();
-          $scope.actionState = "";
+          $scope.actionState = '';
           $scope.clearActionReminder();
           $scope.$digest();
         }
@@ -337,10 +337,10 @@ app.controller("TableController", [
     };
 
     $scope.bet = function () {
-      socket.emit("bet", $scope.betAmount, function (response) {
+      socket.emit('bet', $scope.betAmount, function (response) {
         if (response.success) {
           sounds.playBetSound();
-          $scope.actionState = "";
+          $scope.actionState = '';
           $scope.clearActionReminder();
           $scope.$digest();
         }
@@ -348,10 +348,10 @@ app.controller("TableController", [
     };
 
     $scope.raise = function () {
-      socket.emit("raise", $scope.betAmount, function (response) {
+      socket.emit('raise', $scope.betAmount, function (response) {
         if (response.success) {
           sounds.playRaiseSound();
-          $scope.actionState = "";
+          $scope.actionState = '';
           $scope.clearActionReminder();
           $scope.$digest();
         }
@@ -359,11 +359,11 @@ app.controller("TableController", [
     };
 
     $scope.allIn = function () {
-      if (window.confirm("Andaa neeree you?!")) {
-        socket.emit("allIn", function (response) {
+      if (window.confirm('Andaa neeree you?!')) {
+        socket.emit('allIn', function (response) {
           if (response.success) {
             sounds.playRaiseSound();
-            $scope.actionState = "";
+            $scope.actionState = '';
             $scope.clearActionReminder();
             $scope.$digest();
           }
@@ -402,43 +402,43 @@ app.controller("TableController", [
     };
 
     // When the table data have changed
-    socket.on("table-data", function (data) {
+    socket.on('table-data', function (data) {
       $scope.table = data;
       switch (data.log.action) {
-        case "fold":
+        case 'fold':
           sounds.playFoldSound();
           break;
-        case "check":
+        case 'check':
           sounds.playCheckSound();
           break;
-        case "call":
+        case 'call':
           sounds.playCallSound();
           break;
-        case "bet":
+        case 'bet':
           sounds.playBetSound();
           break;
-        case "raise":
+        case 'raise':
           sounds.playRaiseSound();
           break;
       }
       if (data.log.message) {
-        var messageBox = document.querySelector("#logMessages");
+        var messageBox = document.querySelector('#logMessages');
         var messageElement = angular.element(
-          '<p class="log-message">' + data.log.message + "</p>"
+          '<p class="log-message">' + data.log.message + '</p>'
         );
         angular.element(messageBox).append(messageElement);
         messageBox.scrollTop = messageBox.scrollHeight;
-        if (data.log.notification && data.log.seat !== "") {
+        if (data.log.notification && data.log.seat !== '') {
           if (!$scope.notifications[data.log.seat].message) {
             $scope.notifications[data.log.seat].message = data.log.notification;
             $scope.notifications[data.log.seat].timeout = $timeout(function () {
-              $scope.notifications[data.log.seat].message = "";
+              $scope.notifications[data.log.seat].message = '';
             }, 1000);
           } else {
             $timeout.cancel($scope.notifications[data.log.seat].timeout);
             $scope.notifications[data.log.seat].message = data.log.notification;
             $scope.notifications[data.log.seat].timeout = $timeout(function () {
-              $scope.notifications[data.log.seat].message = "";
+              $scope.notifications[data.log.seat].message = '';
             }, 1000);
           }
         }
@@ -447,37 +447,37 @@ app.controller("TableController", [
     });
 
     // When the game has stopped
-    socket.on("gameStopped", function (data) {
+    socket.on('gameStopped', function (data) {
       $scope.table = data;
-      $scope.actionState = "waiting";
+      $scope.actionState = 'waiting';
       $scope.clearActionReminder();
       $scope.$digest();
     });
 
     // When the player is asked to place the small blind
-    socket.on("postSmallBlind", function (data) {
-      $scope.actionState = "postSmallBlind";
+    socket.on('postSmallBlind', function (data) {
+      $scope.actionState = 'postSmallBlind';
       $scope.setActionReminder();
       $scope.$digest();
     });
 
     // When the player is asked to place the big blind
-    socket.on("postBigBlind", function (data) {
-      $scope.actionState = "postBigBlind";
+    socket.on('postBigBlind', function (data) {
+      $scope.actionState = 'postBigBlind';
       $scope.setActionReminder();
       $scope.$digest();
     });
 
     // When the player is dealt cards
-    socket.on("dealingCards", function (cards) {
-      $scope.myCards[0] = "card_" + cards[0];
-      $scope.myCards[1] = "card_" + cards[1];
+    socket.on('dealingCards', function (cards) {
+      $scope.myCards[0] = 'card_' + cards[0];
+      $scope.myCards[1] = 'card_' + cards[1];
       $scope.$digest();
     });
 
     // When the user is asked to act and the pot was betted
-    socket.on("actBettedPot", function () {
-      $scope.actionState = "actBettedPot";
+    socket.on('actBettedPot', function () {
+      $scope.actionState = 'actBettedPot';
 
       var minRaise =
         $scope.table.lastRaise < $scope.table.bigBlind
@@ -493,8 +493,8 @@ app.controller("TableController", [
     });
 
     // When the user is asked to act and the pot was not betted
-    socket.on("actNotBettedPot", function () {
-      $scope.actionState = "actNotBettedPot";
+    socket.on('actNotBettedPot', function () {
+      $scope.actionState = 'actNotBettedPot';
 
       $scope.betAmount =
         $scope.table.seats[$scope.mySeat].chipsInPlay < $scope.table.bigBlind
@@ -505,8 +505,8 @@ app.controller("TableController", [
     });
 
     // When the user is asked to call an all in
-    socket.on("actOthersAllIn", function () {
-      $scope.actionState = "actOthersAllIn";
+    socket.on('actOthersAllIn', function () {
+      $scope.actionState = 'actOthersAllIn';
 
       $scope.setActionReminder();
       $scope.$digest();
