+++ public/styles/styles.less
@@ -16,8 +16,8 @@ body {
   font-size: 1em;
 }
 
-input[type="text"],
-input[type="number"] {
+input[type='text'],
+input[type='number'] {
   padding: 8px;
   width: 228px;
   border: 1px solid @light-purple;
@@ -37,7 +37,7 @@ input {
 }
 
 button,
-input[type="submit"],
+input[type='submit'],
 .button {
   padding: 16px;
   background-color: rgba(0, 0, 0, 0);
@@ -199,7 +199,7 @@ div.controls {
   }
 }
 
-input[type="range"] {
+input[type='range'] {
   width: 90%;
   margin-top: 4%;
   padding: 0;
@@ -248,14 +248,14 @@ div#chat {
       }
     }
   }
-  input[type="submit"] {
+  input[type='submit'] {
     font-size: 1.3em;
     padding: 5px 8px;
     margin: 1px 0 0 2px;
     background-color: rgba(46, 31, 62, 0.9);
     border-radius: 8px;
   }
-  input[type="text"] {
+  input[type='text'] {
     width: 197px;
     font-size: 1em;
     margin: 3px 0 0 8px;
@@ -354,7 +354,7 @@ div#board-wrap {
 }
 
 .card {
-  background-image: url("/images/deck.png");
+  background-image: url('/images/deck.png');
   background-size: 1040px auto;
   border-radius: 8px 8px 0 0;
   height: 100%;
@@ -759,7 +759,7 @@ div#board-wrap {
       .cell-content {
         padding: 4%;
         padding-left: 8%;
-        input[type="range"] {
+        input[type='range'] {
           margin-top: 0;
         }
         #bet-input {
@@ -821,13 +821,13 @@ div#board-wrap {
     height: 76px;
   }
 
-  div#chat input[type="text"] {
+  div#chat input[type='text'] {
     width: 137px;
     font-size: 1.3em;
     padding: 8px;
   }
 
-  div#chat input[type="submit"] {
+  div#chat input[type='submit'] {
     transform: translate(0px, -2px);
     padding: 7px 8px;
   }
