+++ src/socket-event.js
@@ -1,0 +1,20 @@
+const events = {
+  connection: 'connection',
+  enterRoom: 'enterRoom',
+  leaveRoom: 'leaveRoom',
+  disconnect: 'disconnect',
+  leaveTable: 'leaveTable',
+  register: 'register',
+  sitOnTheTable: 'sitOnTheTable',
+  sitIn: 'sitIn',
+  postBlind: 'postBlind',
+  check: 'check',
+  fold: 'fold',
+  call: 'call',
+  bet: 'bet',
+  raise: 'raise',
+  allIn: 'allIn',
+  sendMessage: 'sendMessage',
+};
+
+module.exports = events;
