+++ src/views/index.jade
@@ -1,0 +1,17 @@
+html(lang="en" ng-app="app")
+	head
+		title Poker
+		link(href="styles/styles.css" rel="stylesheet" type="text/css")
+	body
+		h1 Улаан покерчид
+		#content(ng-view)
+		script(type="text/javascript", src="socket.io/socket.io.js")
+		script(type="text/javascript", src="js/libs/angular.min.js")
+		script(type="text/javascript", src="js/libs/angular-route.min.js")
+		script(type="text/javascript", src="js/app.js")
+		script(type="text/javascript", src="js/controllers/lobby.js")
+		script(type="text/javascript", src="js/controllers/table.js")
+		script(type="text/javascript", src="js/controllers/chat.js")
+		script(type="text/javascript", src="js/services/sounds.js")
+		script(type="text/javascript", src="js/directives/seat.js")
+
